Move commands.go to protocol/header.go. Put Header structs there. Fix types (int -> int16, int32 -> int16) where applicable. Change bit magic to binary.Read/Write. Fix a few lint warnings. in processCommands() removed duplicate code.

+++ internal/crypt/signature/ed25519.go
@@ -25,7 +25,7 @@ func (s *SignatureEd25519) GetName() string {
 	return s.name
 }
 
-func (s *SignatureEd25519) SignLen() int {
+func (s *SignatureEd25519) SignLen() uint16 {
 	return 64
 }
